Add perf based metrics to the system agent
This adds 3 distribution summaries, all recording the values on a per
core basis:

* `sys.cpu.instructionsPerCycle` (num-instructions / num-cycles)
* `sys.cpu.cacheMissRate` (cache-misses / cache-references)
* `sys.cpu.branchMispredictionRate` (branch misses / branch
instructions)

This is done on a global basis, which means the agent needs to be
running as root (`CAP_SYS_ADMIN`), or
`/proc/sys/kernel/perf_event_paranoid` with a value < 1

Note that only a few instance types support the PMC counters. For others the
agent will log a warning at startup indicating that hardware performance
counters are not available.