Deltas and fixed configs for reading perf counters (#27)
The values read from perf are monotonically increasing, so we need to
compute a delta before updating distribution summaries / computing gauge
values.

We ask for total_time_running + total_time_enabled, so we can compute a
scaling factor for the proper value. On my tests the time enabled was
around 2x the time running so the effect is important.