Disable PMU metrics if env variable is set
Due to an issue with an interaction between Xen and the Intel PMUs, in
some cases there can be very high overhead when switching cgroups
around. This PR provides a mechanism that can be used to disable the
collection of `perf` metrics until such problems are resolved.

To disable, just set ATLAS_DISABLE_PMU_METRICS=true as an environment
variable. You should see the following entry in your logs files
indicating this has taken effect:

```
... [info] PMU metrics have been disabled using the env variable ATLAS_DISABLE_PMU_METRICS
```