Gather NTP system metrics
This is a non-daemon specific approach. It gathers metrics from the kernel's
point of view around how the clock is doing. It doesn't capture specific
information like drift, or stability, but it captures precision, and
estimated error.

This can be in turn used to power alerts (if esterror > X, then disable the
node). It's much better than blindly terminating the node.