Add disk latency and percentBusy stats (#56)
Converts the `disk.io.ops` counters to simple timers by adding a
`statistic=totalTime` metric computed as the delta in time
reading/writing so we can compute the average latency for read/write
operations.

In addition, we also add a `disk.percentBusy` time as an
indicator of how much time the disk is being utilized (what `iostat`
calls `%util`). This is the fraction of wall clock time that the device
had at least one request in progress.