Timeout when getting input from external commands
This fixes issue #72: now when an external command is hanging and we
cannot get its output in time we just return an empty string to the
caller.

Note that we cannot simply keep using `popen` under the hood since
`pclose` will wait until the program exits, so it does us no good to
timeout the reads. We need to manually create the pipe and kill the
child when we timeout