Add the ability to tag metrics based on their name (#87)
This can be used by infrastructure teams that are interested in
subscribing to certain system level metrics.

The configuration is done throw a JSON configuration file which includes
a key `"tag_rules"`. This contains an array of rules that specify when
to add custom tags, and the tags to add. The rule can be either an exact
name match (`"name": "net.iface.bytes"`) or a specify they need to start
with a certain prefix. (`"prefix": "sys.load"`). For example:

```
{
  "tag_rules": [
    {
      "prefix": "disk.io.",
      "tags": {
        "tag-key1": "tag-val1",
        "tag-key2": "tag-val2"
      }
    },
    {
      "name": "net.iface.bytes",
      "tags": {
        "tag-key1": "tag-val1",
        "tag-key2": "tag-val2"
      }
    }
  ]
}
```