update build script
A Dockerfile and `build.sh` script have been added to the project, to provide a
reproducible means of running builds locally. To use this build script, you
must provide a BASEOS_IMAGE environment variable, which should resolve to a
reasonable image, such as an Ubuntu LTS version.

The build script accepts one option:

`titus` - Adds a define to enable the Titus code paths.
`shell` - Starts an interactive container session in the source dir.

With no option, it runs the base build.

The bazel output is cached locallly in the `.cache-*` directories, which improves
the speed of rebuilds.

In order to allow the build to complete successfully, it is recommended to
configure Docker Desktop with 8GB RAM.

The `run-build.sh` script is necessary for the Travis build, so it has been
renamed to `travis-build.sh` and a cleanup pass has been done to adjust some
names and quiet some shellcheck warnings.