statically link binutils
The binutils library is included primarily for the purpose of producing
readable stack traces. With dynamic compilation, there was an issue
running the binary against multiple versions of Ubuntu. By switching to
static compilation and building on the oldest target Ubuntu version, we
can produce a binary that will work on all of the versions we support.

Also, use g++-10 in the local development build.