Changes to run as system service (#93)
Titus system services are contained by the executor, no need to contain in the binary. We now run
as a static binary, no need for reexec to lock LD_LIBRARY_PATH and load dylibs. Don't collect perf
metrics when running in system service mode, so no added privileges are needed.