Add TCP memory metric (#96)
In order to be able to track TCP memory usage, we are adding one metric, which
is parsed from `/proc/net/sockstat`:

* net.tcp.memory - Number of bytes allocated to TCP memory.

This metric is a gauge, which will be collected once per minute, with common
tags applied. No additional tags are provided.

The corresponding `/proc/net/sockstat6` does not appear to have any information
on on memory and it looks like it is only used to track the number of `inuse`
sockets for IPv6.