improve exclusion of loopback and ram disk devices
The BaseOS recently changed to move the loopback devices to a different
mount point, which caused them to bypass the existing filtering for the
disk.bytesUsed metrics.

This change uses the device major number as the filter filter pass to
ensure that these devices do not show up in the usage metrics.