disk.io.ops and disk.percentBusy should not report for multiple devices (#108)
Multiple (RAID) devices do not provide the timing data necessary for
calculating the `disk.io.ops` and `disk.percentBusy` statistics. Since
we do not have the data necessary, these metrics should not be reported.

Prior to this change, these timers would not report the `totalTime` and
the `totalOfSquares` components because they were zero increments. The
`count` component would report the number of times an attempt was made
to report the metric and the `max` component was always zero. As a result,
these timers appeared to be counters in Atlas UI.

The `disk.io.bytes` metrics remain relevant for multiple devices.