add ena driver metrics
This change adds support for the following five Elastic Network Adapter
performance metrics, which are available on EC2 instances:

* bw_in_allowance_exceeded
* bw_out_allowance_exceeded
* conntrack_allowance_exceeded
* linklocal_allowance_exceeded
* pps_allowance_exceeded

There are a few apps in our environment which may hit conntrack limits, and
these metrics are a good means of detecting that state. All of these metrics
are configured as Monotonic Counters, since all of the statistics report the
cumulative number of packets queued or dropped on each network interface since
the last driver reset.

The first time that metrics collection occurs, the ethernet interfaces on the
instance will enumerated, so that these statistics may be collected for each
one.