publish `TITUS_NUM_CPU` value as `titus.cpu.requested` (#124)
Titus is moving away from setting the `cgroup.cpu.shares` value. The shares is
not the same as the number of cpus, although it is often incorrectly used that
way.

It is not a good cgroup field to use for a metric, because it is something that
we optionally set, rather than something system-provided, as is the case for
every other cgroup metric. The `cgroup.cpu.processingCapacity` metric is better
in most cases, but some people want to know the requested cpu count.

This change reports the `TITUS_NUM_CPU` environment variable as the
`titus.cpus.requested` metric.