remove cgroupv1 collection methods
All Titus hosts are now running on Jammy with cgroupv2. This change keeps the
cgroupv2 test and boolean flags, to ensure that data collection is valid, and
as a hedge against the development of cgroupv3. Any instances that are not
running with cgroupv2 with will not report any cgroup metrics.