fix sys.cpu.utilization metrics for titus isolate v2
The new 7th generation hardware sets a `cfs_quota` of `max`, which is then read
as a `0` in the number vector. This was breaking utilization calculations.

* Refactor code to remove all remaining traces of cgroup selection, since all
Titus hosts are now on cgroupv2.
* Remove extra `ifdef` macro for `TITUS_SYSTEM_SERVICE` that had no effect.
* The `TITUS_NUM_CPU` environment variable is expected to always be present, and
should always be used to determine the number of CPUs allocated.
* The number of CPUs can be used to determine the value of `cfs_quota`, which
can then be used to calculate available CPU time.
* Refactor metrics collection for cpu time and cpu utilization, to reduce the
number of duplicate calls and simplify the processing.
* Shuffle the order of nvml library initialization, so that we can use the
standard logger to report status.
* Refactor the `setup-venv.sh` script, to make it easier to bootstrap new dev
environments on bionic hosts.