do not set sys.time.estimatedError when time is unsynchronized (#131)
We are trying to gather some data on the accuracy of clock time across the
fleet, and the `sys.time.estimatedError` was reporting a number of 16s values.

These values occur when the kernel has not synchonized the clock yet. This
fact is tracked with the `sys.time.unsynchronized` metric. The `estimatedError`
metric should only have valid results from `ntp_adjtime`.